+++ database/migrations/2021_10_28_215658_create_actual_stocks_table.php
@@ -18,7 +18,7 @@ class CreateActualStocksTable extends Migration
             $table->foreignId('material_id')->constrained();
             $table->foreignId('user_id')->constrained();
             $table->string('batch');
-            $table->bigInteger('quantity');
+            $table->float('quantity');
             $table->timestamps();
             $table->softDeletes();
         });
