+++ database/migrations/2021_11_02_193133_create_product_materials_table.php
@@ -18,8 +18,8 @@ class CreateProductMaterialsTable extends Migration
             $table->foreignId('product_id')->constrained();
             $table->foreignId('material_id')->constrained();
             $table->string('material_uom');
-            $table->bigInteger('material_quantity');
-            $table->bigInteger('product_quantity');
+            $table->float('material_quantity');
+            $table->float('product_quantity');
             $table->softDeletes();
             $table->timestamps();
         });
